Change file structure and import and npm start command for User
