+++ User/helpers/middleware.js
@@ -1,4 +1,4 @@
-const authApi = require('./api/AuthApi.js');
+const authApi = require('#helpers/callsToAuth.js');
 
 const getToken = async (req, res) => {
   try {
